requirements.txt
requests
xmltodict
pymupdf
pytesseract
pillow
pdf2image
python-dotenv
openai
pydantic
fastapi
uvicorn[standard]
streamlit
sentence-transformers
numpy
prometheus-client
azure-storage-blob
azure-search-documents
pytest
httpx
